refactor(cli): dissolve util/awareness.ts; AwarenessState lives next to LoadedWorkspace
The file had collapsed to 3 lines of meaningful code — one type alias and
a one-line readPeers wrapper. The type origin moves to load-config.ts
(adjacent to LoadedWorkspace, which already references the typed
Awareness wrapper); readPeers inlines at its 2 call sites:

  workspace.awareness?.peers() ?? new Map<number, AwarenessState>()

One less file to navigate, no abstraction lost — peers() already filters
self and validates against the schema.
